api/requirements.txt
aiohttp

gunicorn
pillow
requests
SQLAlchemy
